Fix scrolling when sending a message
We're listening to new messages come in, but for that to happen they
need to be saved to Firebase first. Before any of that could happen we'd
already sent out the call to scroll to bottom, so the app scrolled to
the current bottom, then the listener called out, added the new message
but we don't scroll futher.

The solution is quite involved, but something that I've been meaning to
do for a while. Before messages are persisted now, they're stored
locally in Redux. If you try sending a message after this is merged,
you'll see that the message bubble is at half opacity at first-that's to
denote that's it's "local only". When it goes to full opacity, that's
when it's been persisted.