Fix StoryCard's staying the same size when resizing
This worked differently from what I originally envisioned. It turns out
the Redux `connect` higher-order component has a shouldComponentUpdate
that blocks updates if the Redux state doesn't change. That means just
re-rendering the whole app on resize doesn't actually re-render the
StoryMaster because the redux state didn't change.

This works around that by just doing it in the StoryMaster itself
:shrugs: