Begin an overall refactor of the Right Column...
The Right Column index file has become a beast as we've been shipping rapidly over the last few weeks. With the addition of DMs, I was trying to reuse the same component container to handle entirely different behavior and user interactions. This made the component even more massive and confusing.

This commit builds upon the existing DM branch to bring some sanity to the Right Column component, while also addressing underlying blockers to the DM feature (like weird re-rendering bugs when people send messages to stories when I was viewing a DM thread...)