Fix Iris restarting for changes in src/
The code splitting PR introduced a new manifest file
(`react-lodable.json`) that's necessary for SSR + code splitting and is
emitted to the `build/` folder. Unfortunately, iris listened to changes
to that file as it requires it, meaning every time you changed the
frontend that file would be changed, then iris would restart, breaking
the development experience.

Fixed by adding build/ to the watch ignore plugin.