Hotfix web push notifications
This explains why folks weren't getting any web push notifications, the
signup code was broken somehow. Their keys in the db would show up as:

```
{
  keys: {
    auth: '[object Object]',
    p256dh: '[object Object]'
  }
}
```

This patch fixes it. We should check in the db how many people were
affected by this and tell them to subscribe again after this is
deployed.