Add SEO enhancements to infinite loading thread messages
Google does not index infinite loading content, obviously their spider
doesn't scroll the page. This is a problem for us since we now infinite
paginate our thread messages, i.e. Google only ever sees the top 50
messages at most.

This patch introduces a `?msgsafter` and `?msgsbefore` query param to
thread views, which paginates the messages based on the cursors provided
to them. On top of that, we also render [a `link[rel="prev"]` and/or a
`link[rel="next"]`](https://webmasters.googleblog.com/2011/09/pagination-with-relnext-and-relprev.html) to indicate to Google that there's more content on that page.

These `<link>`s use the `?msgsafter` and `?msgsbefore` query params to
load the next/previous pages of the messages. This should hopefully lead
to Google fully indexing us again. :tada:

Eventhough we want Google to _index_ all the messages, we never ever
want Google to point users to a page that has `?msgsafter` or
`?msgsbefore`. That'd be a horrible experience as a user, since you
usually want to start at the top rather than somewhere in the middle.

That's why I also include a `link[rel="canonical"]` whenever either of
the two other `<link>`s is rendered.

> Note: I don't know if we should always render a `rel="canonical"` even
when those query params aren't specified. I don't think that's a smart
default given in how many places a thread could be rendered, which might
confuse Google a ton given there could be a rel="canonical" pointing to
`/thread/asdf123` on `/spectrum/general` because the thread slider is
open.

> This seems like a safe bet, worst case scenario people just see the
second page of the messages because that's where Google saw their search
query. Not too bad.

I'll look into doing this for thread feeds on community and channel
views next, given how important it is for us to have all threads indexed
but also how easy this was to implement.

Ref issue: #1533