Create a specific changefeed per user/subscription
This is based on #2433.

It reverts the second change made on that branch (6522378) so that we
have one db changefeed per user and subscription again.

Previously we would have one changefeed listening to all new
messages in the messages table and filter them after the fact.

This patch implements a changfeed per user and subscription,
where the changefeed is automatically filtered to only include
the changes we care about. (in this case per thread)