Track userThreadLastSeen properly from subscriptions
Previously that code would run everytime a new message was sent, thusly
setting `userThreadLastSeen` a million times. That's really unnecessary,
the only thing we really want to change that is once a user
unsubscribes. (i.e. once they click away from the thread)

This patch is based on #2509, which adds that ability to the `asyncify`
util. This tiny change should make lastSeen even more useful!