Fix thread chatinput ssr rehydration issue
The issue was that on the server, `storedContent` was always null, but
on the client it might've existed if I previously wrote a draft message.

We can remove this case entirely, since it doesn't make sense to render
a chat input if I'm not a member—even if I have previous stored content.
By removing that condition server and client have to match again!