Fix broken authentication
The ServiceWorker would automatically return our app shell for all
navigation requests, including to `/auth/xyz`—which meant that users
were served the app instead of getting redirected to their auth provider
of choice.

This fixes it by disabling the built-in `appShell` option, copying
exactly what it does but filtering out the `/auth` route. (and the
`/api` route for good measure)