Fix forwarding to newest version of JS bundle chunks
ServiceWorker installs are failing because old ServiceWorkers are trying
to cache old versions of the bundle. Unfortunately, the hash changes with
each release so if the SW tries to cache `UserSettings.asdf123.chunk.js`
and we've released a new verion, that file is called something entirely
different like `UserSettings.lkjli53.chunk.js` instead!

This fixes the issue by forwarding old asset names to the new name, i.e.
if a user requests `UserSettings.asdf123.chunk.js` but the current name
is `UserSettingsings.lkjli53.chunk.js` we forward them to the new version.

This should potentially fix some ServiceWorker problems!