Add *AndRole indexes to usersChannels
This adds two new indexes to `usersChannels` and rewrites all queries to
use them:

- `userIdAndRole`
- `channelIdAndRole`

"Role" is one of the roles a user can have in a channel, and is based on
the `is{Role}` fields, i.e. "Role" can be one of "blocked", "pending",
"owner", "moderator", "owner".

Note that the index prefers roles in that order, so even if a write
query is inconsistent and a member has `isMember: true` but also
`isBlocked: true` the user will be counted as blocked.

----

This is modeled after the same thing we did for `usersCommunities`, but
I realized that adding 5 indexes to a table as big as `usersCommunities`
and `usersChannels` is pretty expensive storage-wise. (I could've also
implemented the same indexes as `usersCommunities`, i.e.
`userIdAndIsMember`, `userIdAndIsBlocked`, etc.)

This achieves the same thing and makes all queries be indexed, but
requires a lot less storage. We should probably also move the
usersCommunities indexes to this `*AndRole` system...